requirements.txt
av
torch
torchvision
pims
scipy
numpy
matplotlib
opencv_python
tqdm
Pillow
h5py
